refactor(atomic): move results list components to separate folder
https://coveord.atlassian.net/browse/KIT-1517
