+++ packages/atomic/src/components/atomic-result-table-placeholder/atomic-result-table-placeholder.pcss
@@ -1,5 +1,5 @@
 @import '../../global/global.pcss';
-@import '../atomic-result-list/atomic-result-table.pcss';
+@import '../result-lists/atomic-result-list/atomic-result-table.pcss';
 
 :host {
   display: grid;
