+++ packages/atomic/src/components/result-lists/atomic-folded-result-list/atomic-folded-result-list.pcss
