+++ packages/atomic/src/components/result-lists/atomic-result-list/adopted-stylesheets.d.ts
