+++ packages/atomic/src/components/result-lists/atomic-result-list/atomic-result-table.pcss
